pyproject.toml
[project]
name = "nyc-curb-utilization-tool"
version = "0.1.0"
description = "NYC-focused curb utilization analysis for parking legality, lane obstruction, and occupancy insights."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "pyyaml>=6.0.2",
  "streamlit>=1.38.0",
  "pandas>=2.2.2",
  "ultralytics>=8.2.0",
  "opencv-python-headless>=4.9.0",
  "shapely>=2.0.0",
  "Pillow>=10.0.0",
  "plotly>=5.22.0",
  "streamlit-drawable-canvas>=0.9.3",
  "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
